Huge Update
This update includes major changes to the game state, game mode, and game instance.  A lot of UI work has been done for graphics settings and the settings menu in general.  There was also and update to bring in the last enemy type which was taken from Valley of the Ancients.